Move read_log_file_lines() to test helpers module.
